Added ability to sync to a file on the filesystem called scratch.txt
Sync is only one way right now. Need to come up with either a watch, or
a polling mechanism to get the sync to go back the other way. Need to figure
out a strategy to key the input that is on the webpage to the text currently
being displayed in the scratch pad because I don't want to mistakenly overwrite
what's inside the browser. It whould only sync if the correct text is being
displayed in the browser.